+++ lib/goodcheck/config_loader.rb
@@ -240,8 +240,8 @@ module Goodcheck
     def load_import(rules, import)
       Goodcheck.logger.info "Importing rules from #{import}"
 
-      import_loader.load(import) do |content|
-        json = JSON.parse(JSON.dump(YAML.safe_load(content, filename: import)), symbolize_names: true)
+      import_loader.load(import) do |content, filename|
+        json = JSON.parse(JSON.dump(YAML.safe_load(content, filename: filename)), symbolize_names: true)
 
         Schema.rules.coerce json
         load_rules(rules, json)